Add Example Files
This adds a new directory for examples. Currently we have three, one for each solution method. Python is also complaining about your last name having a non ascii character. I took it out, but you can of course add it back-just define a character encoding scheme and it shouldn't complain! Also removed the last edited date because we have all of that info in github.